Moved some functions to shared, since they are common
